Got the scrolling to work
Trying to make the sidebar sit above the profile division at the bottom of the page when the user scrolls down. I was trying to do that by detecting when the user has scrolled down to the bottom of the forum part.

Also still having problems deleting elements from the navbar